requirements.txt
cloudscraper
python-dotenv
azure-identity
msgraph-sdk
gpiozero